requirements.txt
# Strands Agents
strands-agents>=1.0.0

# Slack SDK
slack-sdk>=3.35.0
